fix: Completion item label spacing for non "function-like" items
## Issue

When dealing with completion items the default
`completion.menu.draw.components.label.text(ctx)` function just returns
the concatenation of `ctx.label` and `ctx.label_detail` which creates
scenarios such as "MyAwesomeModinterface" which is wrongly formatted.

## Solution

The solution consists in checking the `ctx.kind` against the internal
`types.CompletionItemKind` table and for any kind that is not either
`Method`, `Function` or `Constructor` we add a space between the
`ctx.label` and `ctx.label_detail`.