fix: documentation losing syntax highlighting on doc reopen (#768)
Issue
=====

Providers (completion plugins) can provide a custom implementation for
rendering the documentation. Some need to use e.g. Neovim's built-in
regex based syntax highlighting. If the documentation window is closed
and reopened, the syntax highlighting is lost.

Solution
========

Allow plugins to redraw the documentation window when the documentation
window is reopened.

Closes #703