doc: use the correct way to check comment block
Using node:type() do not work as expected, sometimes, the type will go
with a "chunk" or even empty string, we must iterate the node of current
line to check if current line is inside a comment block. Besides, in the
code, there is no `ctx` passed to the default function, just remove the
parameter in the doc.