fix(snippet): correctly expand luasnip hidden snippet
Commit #f0f68a introduced a change in the `snippet_forward` behaviour
that is not expected and semantically incorrect. It expands a snippet.

The initial intention was to handle hidden snippet but it was missing
additional checks; the completion menu must no be visible and must no be
already in a snippet.

The behaviour should be more viable with this change.

Related #1503
Related #1515