fix: disable auto brackets by default for cpp filetype (#1595)
* fix: disable kind_resolution by default for cpp filetype

For auto_brackets.kind_resolution, add 'cpp' to the blocked_filetypes
list. This is necessary to ensure appropriate completion behavior, as
there are some completion contexts (such as using-declarations) where it
is not appropriate to insert parenthesis when completing a function
name.

Also, sync documentation of auto_brackets.kind_resolution so that it is
consistent with the defaults in accept.lua.

* fix: disable all auto brackets for cpp

---------

Co-authored-by: Liam Dyer <liamcdyer@gmail.com>