fix: serialization issue with draw function (#1719)
Issue
=====

When using the rust implementation for fuzzy matching, when the frecency
database is being updated after accepting a completion, the completion
item is serialized and passed to the rust code.

If the completion item has a `draw` function (used for a custom
implementation for drawing the documentation), it will throw an error
when serialized with `vim.mpack.encode()`:

> ...ta/nvim-data/lazy/blink.cmp/lua/blink/cmp/fuzzy/init.lua:44: can't
> serialize object of type 6

A reproduction is available here:
https://github.com/mikavilpas/blink-ripgrep.nvim/issues/185#issuecomment-2759181172

Solution
========

Hide the `draw` function from the completion item before serializing it
to avoid the error.