feat(completion.trigger): add "show_on_backspace" option
As of now, the completion menu will not display when the user backspaces
into a completion they accepted.

With "completion.trigger.show_on_backspace = true", the user can have
the completion menu display when they backspace into a keyword or
trigger character (that isn't a space) that they have just
accepted. This intends to duplicate nvim-cmp's default behavior and
shares characteristics with VSCode. This option is opt-in.