feat(completion.trigger): add `show_on_backspace_*` options (#1813)
* feat(completion.trigger): add "show_on_backspace" option

As of now, the completion menu will not display when the user backspaces
into a completion they accepted.

With "completion.trigger.show_on_backspace = true", the user can have
the completion menu display when they backspace into a keyword or
trigger character (that isn't a space) that they have just
accepted. This intends to duplicate nvim-cmp's default behavior and
shares characteristics with VSCode. This option is opt-in.

* feat: use `vim.on_key` for backspace detection

* feat(trigger): opt to trigger menu when bs after insert enter or accept

- Add the "show_on_backspace_after_accept" boolean option, to show the completion window after accepting a completion and then backspacing into a keyword. This is enabled by default.

- Add the "show_on_backspace_after_insert_enter" boolean option, to show the completion window after entering insert mode and backspacing into keyword. This is enabled by default.

To make these features possible, add
    - mechanism to create and consume state, in which completion
      acceptance and InsertEnter events are found and stored for one
      movement
    - an event listener for completion acceptance using the existing
      event listener infrastructure
    - a check for the current event in the make_cursor_moved function

NOTICE: The "show_on_backspace_after_insert_enter" will not work on first InsertEnter if blink is lazy-loaded on that event

* feat: localize backspace state

* feat: `show_on_backspace`, `show_on_backspace_in_keyword` options, clear state on insert leave

* docs: add new options to reference

* chore: drop unused requires

---------

Co-authored-by: Liam Dyer <liamcdyer@gmail.com>